User mgmt ui (#81)
This will allow us to manage the Alert Manager users via the UI. We had to allow CRUD of DB users in the backend API. So this will DEPRECATE the need for putting the users in the Nomad config file.

This also forces the default Alerts view to show only the users team alerts. The user can still switch to another users view, copy and paste the query. But otherwise it will default to the users team view. You can no longer view ALL teams alerts. However, if this is necessary we can add in that option.