Add login warnings and help text about might have failed (#82)
When the user fails to login they are not given any indication why it failed. This diff adds a notification with helpful texts. Especially now that we require a username to be added to alert manager DB, this let's the user know if that is why they cannot login.

It also refactor the api call so the "client" is responsible for handling the different errors, instead of the API lib asserting how it should fail.

The following login errors are caught:

    Cannot connect to ldap servers, most likely VPN issues. It alerts the user to check vpn
    Incorrect username/password
    User is not added to the alert manager DB